backend/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.1
Flask-Migrate==4.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
PyJWT==2.12.1
weasyprint==68.1
python-docx==1.1.2
openpyxl==3.1.5
